Checkpoint 4
Tigera secrets add

TODO: Testing needs to continue on an environment with dual tor
calico is failing to reach ready due to the second ToR switch being
unconfigured in the test environment.

More work needs to be done on the testing environment terraform to
have the second tor configured correctly.